pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scaruffi-tidal"
version = "1.0.0"
description = "Create Tidal playlists from Scaruffi's classical music recommendations"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pyyaml>=6.0",
    "discogs-client>=2.3.0",
    "tidalapi>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
scaruffi-tidal = "scaruffi_tidal.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["domain*", "application*", "infrastructure*", "scaruffi_tidal*"]
